Moved EnumerableExtensions into System.Collections.Generic
