+++ src/EnumerableExtensions/DictionaryExtensions.cs
@@ -3,6 +3,7 @@ using InsightArchitectures.Utilities;
 // ReSharper disable once CheckNamespace
 namespace System.Collections.Generic;
 
+#pragma warning disable SA1649
 /// <summary>
 /// A set of extension methods on the <see cref="IDictionary{TKey,TValue}"/> type.
 /// </summary>
@@ -54,3 +55,4 @@ public static class InsightArchitecturesDictionaryExtensions
         }
     }
 }
+#pragma warning restore SA1649
